Extraction of text from image implemented
For this type of extraction pytesseract is used, which itself uses
tesseract-orc. A new method textFromImage() has been defined. It takes
the path to the image file, and a selected language as arguments. Both
inputs are validated, before they are used to extract the text from the
image. The result is then returned. I a next step the translation of
the extracted text is planed.