Translation and language detection
Translation and language detection use goslate. This is only temporary,
and a future implementation should use Bing translate. The extracted
text is translated into a specified target language from a source
language. The source language can be specified or not. In a next step
the information will be combined in a single JSON object.