Definition of result in form of JSON
The result is returned in form of a JSON with the original text,
translated text, source language, target language and detected
language. Enrichment of the data is implemented at a later point. The
next step is going to be a definition of different startup parameters
for the server, as well as documentation of the written code.