Detect text again, Docker and Documentation
If no source language is specified the application will try to extract
text from the image using tesseract, detect the language using goslate,
and then try to detect text again using tesseract. This is supposed to
minimize the errors made by the ocr engine, especially then when a
language with a lot of special characters is used.
Methods changed to static.
Added license, documentation, readme and a Dockerfile.